pyproject.toml
[project]
name = "moc-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "httpx>=0.28.1",
    "ipython>=9.12.0",
    "lightrag-hku[ollama]>=1.4.15",
    "ollama>=0.6.1",
    "openai>=2.32.0",
    "openpyxl>=3.1.5",
    "psycopg2-binary>=2.9.12",
    "pymupdf>=1.27.2.2",
    "python-dotenv>=1.2.2",
]
